requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
yfinance>=0.2.18
PyPDF2>=3.0.0
matplotlib>=3.8.0
plotly>=5.20.0
streamlit>=1.35.0
pytest>=7.4.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
